Refactoring Home & Details
