+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@formkit/auto-animate": "^1.0.0-beta.3",
     "axios": "^0.27.2",
     "bootstrap": "^5.2.1",
     "jquery": "^3.6.1",
@@ -10,8 +11,9 @@
     "react": "^18.2.0",
     "react-bootstrap": "^2.5.0",
     "react-dom": "^18.2.0",
-    "react-dropdown": "^1.10.0",
     "react-icons": "^4.4.0",
+    "react-lazy-load-image-component": "^1.5.5",
+    "react-loader-spinner": "^5.3.4",
     "react-router-dom": "^6.4.0",
     "react-scripts": "5.0.1",
     "styled-components": "^5.3.5",
